added the specific example code (our code is VisionOld)
